Day 4
Need to create separate functions one, that will run the table creation and then add rows to each but make it so that the secondary keys are being filled by the primary keys and also make and update function that allows me to populate each table while using each tables previous information. Recommend looking at the Tim guy's video for a better understanding of connecting the tables.